feat: 기존에 만들어 놓은 api에 맴버정보 연동

+++ src/main/java/com/h10/sideproject/poll/service/PollService.java
@@ -2,6 +2,8 @@ package com.h10.sideproject.poll.service;
 
 import com.h10.sideproject.category.entity.Category;
 import com.h10.sideproject.category.repository.CategoryRepository;
+import com.h10.sideproject.member.entity.Member;
+import com.h10.sideproject.member.repository.MemberRepository;
 import com.h10.sideproject.poll.dto.PollRequestDto;
 import com.h10.sideproject.poll.dto.PollResponseDto;
 import com.h10.sideproject.poll.entity.Poll;
@@ -9,19 +11,23 @@ import com.h10.sideproject.poll.repository.PollRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
 public class PollService {
+    private final MemberRepository memberRepository;
+
     private final PollRepository pollRepository;
 
     private final CategoryRepository categoryRepository;
 
     @Transactional
-    public ResponseEntity<?> createPoll(PollRequestDto pollRequestDto) {
+    public ResponseEntity<?> createPoll(PollRequestDto pollRequestDto, UserDetails user) {
         Category category = categoryRepository.findByName(pollRequestDto.getCategory()).orElse(null);
+        Member member = memberRepository.findByEmail(user.getUsername()).orElse(null);
         if(category == null){
             category = categoryRepository.save(Category.builder().name(pollRequestDto.getCategory()).build());
         }
@@ -34,30 +40,37 @@ public class PollService {
                 .choice2(pollRequestDto.getChoice2())
                 .choice2_img(pollRequestDto.getChoice2_img())
                 .view(0L)
+                .member(member)
                 .build()
         );
         return new ResponseEntity<>("설문 작성 완료",HttpStatus.OK);
     }
 
     @Transactional
-    public ResponseEntity<?>updatePoll(PollRequestDto pollRequestDto, Long poll_id) {
+    public ResponseEntity<?>updatePoll(PollRequestDto pollRequestDto, Long poll_id, UserDetails user) {
         Poll poll = pollRepository.findById(poll_id).orElse(null);
-        if(poll != null){
+        Member member = memberRepository.findByEmail(user.getUsername()).orElse(null);
+        if(poll != null && poll.getMember().getId() == member.getId()){
             Category category = categoryRepository.findByName(pollRequestDto.getCategory()).orElse(null);
             if(category == null){
                 category = categoryRepository.save(Category.builder().name(pollRequestDto.getCategory()).build());
             }
             poll.update(pollRequestDto,category);
+            return new ResponseEntity<>("설문 수정 성공",HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>("권한 없음",HttpStatus.OK);
         }
-        return new ResponseEntity<>("설문 수정 성공",HttpStatus.OK);
     }
 
-    public ResponseEntity<?> deletePoll(Long poll_id) {
+    public ResponseEntity<?> deletePoll(Long poll_id, UserDetails user) {
         Poll poll = pollRepository.findById(poll_id).orElse(null);
-        if(poll != null){
+        Member member = memberRepository.findByEmail(user.getUsername()).orElse(null);
+        if(poll != null && poll.getMember().getId() == member.getId()){
             pollRepository.deleteById(poll_id);
+            return new ResponseEntity<>("설문 삭제 성공",HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>("권한 없음",HttpStatus.OK);
         }
-        return new ResponseEntity<>("설문 삭제 성공",HttpStatus.OK);
     }
 
     @Transactional
@@ -65,6 +78,7 @@ public class PollService {
         Poll poll = pollRepository.findById(poll_id).orElse(null);
         poll.plusView();
         PollResponseDto pollResponseDto = PollResponseDto.builder()
+                .nickname(poll.getMember().getNickname())
                 .title(poll.getTitle())
                 .category(poll.getCategory().getName())
                 .choice1(poll.getChoice1())
